+++ libs/gesha-core/src/testing/mod.rs
@@ -4,7 +4,7 @@ mod test_case;
 pub use test_case::TestCase;
 
 mod test_case_map;
-pub use test_case_map::TestCaseMap;
+pub use test_case_map::{run_parallel, TestCaseMap};
 
 mod test_suite;
 pub use test_suite::TestSuite;
