+++ libs/gesha-core/src/testing/test_case_map.rs
@@ -1,7 +1,9 @@
 use crate::testing::TestCase;
-use crate::Error;
+use crate::{Error, Result};
 use std::collections::HashMap;
-use tokio::task::{Id, JoinError};
+use std::future::Future;
+use tokio::task::{Id, JoinError, JoinHandle};
+use tracing::Instrument;
 
 #[derive(Default)]
 pub struct TestCaseMap<A>(HashMap<Id, TestCase<A>>);
@@ -16,7 +18,7 @@ impl<A> TestCaseMap<A> {
         self
     }
 
-    pub fn extract(&mut self, id: Id) -> crate::Result<TestCase<A>> {
+    pub fn extract(&mut self, id: Id) -> Result<TestCase<A>> {
         self.0
             .remove(&id)
             .ok_or_else(|| Error::ThreadNotFound(id.to_string()))
@@ -30,7 +32,7 @@ impl<A> TestCaseMap<A> {
         (handles, map.push(id, case))
     }
 
-    pub fn flatten<B>(&mut self, result: Result<crate::Result<B>, JoinError>) -> crate::Result<B> {
+    pub fn flatten<B>(&mut self, result: std::result::Result<Result<B>, JoinError>) -> Result<B> {
         match result {
             Ok(x) => x,
             Err(cause) => Err(Error::JoinError {
@@ -39,4 +41,23 @@ impl<A> TestCaseMap<A> {
             }),
         }
     }
+}
+
+pub fn run_parallel<X, Y, F, Fut>(
+    xs: Vec<TestCase<X>>,
+    f: F,
+) -> (Vec<JoinHandle<Result<Y>>>, TestCaseMap<X>)
+where
+    F: Fn(TestCase<X>) -> Fut,
+    Fut: Future<Output = Result<Y>> + Send + 'static,
+    X: Clone,
+    Y: Send + 'static,
+{
+    xs.into_iter()
+        .map(|x| {
+            let cloned = x.clone();
+            let handle = tokio::spawn(f(x).in_current_span());
+            (handle.id(), cloned, handle)
+        })
+        .fold((vec![], TestCaseMap::new()), TestCaseMap::accumulate)
 }
